Revert colspan to integer

+++ app/src/main/java/se/nullable/flickboard/model/KeyM.kt
@@ -74,7 +74,7 @@ data class KeyM(
     // Fast actions are performed immediately when detected, rather than when the finger is released.
     val fastActions: Map<Direction, Action> = mapOf(),
     val holdAction: Action? = null,
-    val colspan: Float = 1F,
+    val colspan: Int = 1,
     val fixedWidth: Dp? = null,
     val rendered: Boolean = true,
     val shift: KeyM? = null,
